requirements.txt
Flask
python-pptx
Pillow
requests
werkzeug